# PERF: do nanargmin/max in-order for integer dtypes